Move the contentful generated types to a folder under src
