+++ src/types/TypeComponentRichImage.ts
@@ -9,8 +9,8 @@ export interface TypeComponentRichImageFields {
 
 export type TypeComponentRichImageSkeleton = EntrySkeletonType<TypeComponentRichImageFields, "componentRichImage">;
 export type TypeComponentRichImage<Modifiers extends ChainModifiers, Locales extends LocaleCode> = Entry<TypeComponentRichImageSkeleton, Modifiers, Locales>;
-export type TypeComponentRichImageWithoutLinkResolutionResponse = TypeComponentRichImage<"WITHOUT_LINK_RESOLUTION">;
-export type TypeComponentRichImageWithoutUnresolvableLinksResponse = TypeComponentRichImage<"WITHOUT_UNRESOLVABLE_LINKS">;
+export type TypeComponentRichImageWithoutLinkResolutionResponse = TypeComponentRichImage<"WITHOUT_LINK_RESOLUTION", string>;
+export type TypeComponentRichImageWithoutUnresolvableLinksResponse = TypeComponentRichImage<"WITHOUT_UNRESOLVABLE_LINKS", string>;
 export type TypeComponentRichImageWithAllLocalesResponse<Locales extends LocaleCode = LocaleCode> = TypeComponentRichImage<"WITH_ALL_LOCALES", Locales>;
 export type TypeComponentRichImageWithAllLocalesAndWithoutLinkResolutionResponse<Locales extends LocaleCode = LocaleCode> = TypeComponentRichImage<"WITHOUT_LINK_RESOLUTION" | "WITH_ALL_LOCALES", Locales>;
 export type TypeComponentRichImageWithAllLocalesAndWithoutUnresolvableLinksResponse<Locales extends LocaleCode = LocaleCode> = TypeComponentRichImage<"WITHOUT_UNRESOLVABLE_LINKS" | "WITH_ALL_LOCALES", Locales>;
